// licentainangular.client/src/app/models/NewDTOs.ts
export interface AdresaNestedDTO {
  Bloc?: string;
  Scara?: string;
  Etaj?: string;
  Apartament?: string;
  Strazi: StraziDTO; // Observă că aici 'Strazi' este un obiect
}

export interface StraziDTO {
  DenumireStrada: string;
  Nr: number;
  Localitati: LocalitatiDTO;
}

export interface LocalitatiDTO {
  DenumireLocalitate: string;
  Judete: JudeteDTO;
}

export interface JudeteDTO {
  DenumireJudet: string;
  Tari: TariDTO;
}

export interface TariDTO {
  DenumireTara: string;
}


export interface UserAddressDTO {
  idUser: number;
  idAdresa: number;
}